Adds proper date object instead of string
